Removed contacts and added schema for signUp table and api_routes.
