+++ models/SignUp.js
@@ -1,0 +1,28 @@
+// Creates table SignUp in database
+
+module.exports = (sequelize, DataTypes) => {
+
+    const SignUp = sequelize.define("SignUp", {
+
+        username: {
+            type: DataTypes.STRING,
+            allowNull: false,
+            unique: true,
+        },
+        email: {
+            type: DataTypes.STRING,
+            allowNull: false,
+            unique: true,
+        },
+        password: {
+            type: DataTypes.STRING,
+            allowNull: false,
+        },
+        confirmPassword: {
+            type: DataTypes.STRING,
+            allowNull: false,
+        },
+    });
+
+    return SignUp;
+};
