+++ server.js
@@ -18,12 +18,11 @@ const db = require("./models");
 const userRouter = require("./api_routes/apiUser");
 app.use("/user", userRouter);
 
-const contactRouter = require("./api_routes/apiContacts");
-app.use("/contacts", contactRouter);
-
 const loginRouter = require("./api_routes/apiLogin");
 app.use("/login", loginRouter);
 
+const signUpRouter = require("./api_routes/apiSignUp");
+app.use("/signup", signUpRouter);
 
 db.sequelize.sync().then(() => {
     app.listen(PORT, () => {
